backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
langchain>=0.1.0
langchain-openai
langchain-community
openai
faiss-cpu
tiktoken
passlib[bcrypt]
python-jose
